Added parameter less constructor to Card
Changed deck generation method to add a UniqueIdentifier for each type of playing card (Red1) is unique, but will share the same identifier with the 2nd red1 card.

This identifier is used in the new player card sorting method in
Player.cs.

Player CardSorting is now called when the player changes, when a player discards a card or picks up a card.